+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/impl/MessageService.java
@@ -10,7 +10,7 @@ public class MessageService {
  
     @Autowired
     private MessageSource messageSource;
- 
+    
     public String getMessage(String code) {
         return messageSource.getMessage(code, null, LocaleContextHolder.getLocale());
     }