+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/impl/UserServiceImpl.java
@@ -13,6 +13,7 @@ import com.tanaguru.service.MailService;
 import com.tanaguru.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.mail.MailException;
 import org.springframework.stereotype.Service;
 
 import java.text.DateFormat;
@@ -159,10 +160,21 @@ public class UserServiceImpl implements UserService {
                                 .append(longDateFormat.format(attempt.getBlockedUntil()));
                     }
                 }
-                if(sendAdminMail) {
-                    mailService.sendSimpleMessage(ADMIN_MAIL,ATTEMPTS_MAIL_SUBJECT_ADMIN, builder.toString());
+                try {
+                    if(sendAdminMail) {
+                        mailService.sendSimpleMessage(ADMIN_MAIL,ATTEMPTS_MAIL_SUBJECT_ADMIN, builder.toString());
+                        LOGGER.info("[User {}] account blocking email sent to admin", user.getId());
+                    }
+                }catch(MailException e) {
+                    LOGGER.error("[User {}] Unable to send the account blocking email to admin", user.getId());
                 }
-                mailService.sendSimpleMessage(user.getEmail(), ATTEMPTS_MAIL_SUBJECT_USER, builder.toString());
+                try {
+                    mailService.sendSimpleMessage(user.getEmail(), ATTEMPTS_MAIL_SUBJECT_USER, builder.toString());
+                    LOGGER.info("[User {}] account blocking email sent to user", user.getId());
+                }catch(MailException e) {
+                    LOGGER.error("[User {}] Unable to send the account blocking email to user", user.getId());
+                }
+                
                 break;
 
             default: //Do nothing