+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/service/impl/runner/AuditRequestConsumerServiceImpl.java
@@ -7,7 +7,10 @@ import com.tanaguru.domain.exception.CustomEntityNotFoundException;
 import com.tanaguru.repository.*;
 import com.tanaguru.runner.factory.AuditRunnerFactory;
 import com.tanaguru.service.AuditService;
+import com.tanaguru.service.MailService;
 import com.tanaguru.service.ResultAnalyzerService;
+import com.tanaguru.service.impl.MessageService;
+
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
@@ -55,7 +58,11 @@ public class AuditRequestConsumerServiceImpl extends AuditRequestServiceSyncStan
             ResultAnalyzerService resultAnalyzerService,
             TestHierarchyRepository testHierarchyRepository,
             ElementResultRepository elementResultRepository,
-            AuditReferenceRepository auditReferenceRepository) {
+            AuditReferenceRepository auditReferenceRepository,
+            MailService mailService,
+            MessageService messageService,
+            ActRepository actRepository,
+            ContractUserRepository contractUserRepository) {
 
         super(pageRepository,
                 auditRepository,
@@ -67,7 +74,11 @@ public class AuditRequestConsumerServiceImpl extends AuditRequestServiceSyncStan
                 resultAnalyzerService,
                 testHierarchyRepository,
                 elementResultRepository,
-                auditReferenceRepository);
+                auditReferenceRepository,
+                mailService,
+                messageService,
+                actRepository,
+                contractUserRepository);
         this.auditRequestConsumer = auditRequestConsumer;
     }