+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - At launch we link the repositories present with the version of the webext engine used
 - Endpoint healthcheck , new key in common properties : management.endpoint.health.show-details, in order to show details or not
 - JSON export of audits and pages
+- Send email at the end of the audit (site, scenario, group of pages)